Add cofirmation button on delete user.

+++ app/asset/js/main.js
@@ -11,11 +11,31 @@ $(document).ready(function() {
 function confirmDeleteUser(id_user) {
     var href = '/app/user/deleteUser.php';
     //window.location = href;
-    $.ajax({
-        type: 'DELETE',
-        url: href + '/?id_user='+id_user,
-        success: function (result) {
-            window.location = '/app/user/listUser.php';
-        }
+    $.confirm({
+      title: 'Confirmation!',
+      content: 'Vous voulez vraiment supprimer?',
+      icon: 'fa fa-warning',
+      theme: 'material',
+      closeIcon: true,
+      animation: 'scale',
+      type: 'red',
+      columnClass: 'col s4 offset-s4',
+      buttons: {
+          valider: {
+              text: 'OK',
+              btnClass: 'btn-red',
+              action: function () {
+                  $.ajax({
+                      type: 'DELETE',
+                      url: href + '/?id_user='+id_user,
+                      success: function (result) {
+                          window.location = '/app/user/listUser.php';
+                      }
+                  });
+              }
+          },
+          cancel: function () {
+          },
+      }
     });
 }
